feat: moved faq dropdown into a separate component
